# openspec/openapi-v1.yaml
openapi: 3.1.0
info:
  title: AstrBot OpenAPI v1
  version: 0.1.0
  description: >
    Target REST contract for migrating AstrBot HTTP APIs to /api/v1.
    Dynamic AstrBot configuration payloads are intentionally modeled as open
    JSON objects because their schemas are provided at runtime by template
    endpoints.
    Developer API keys currently support these scopes only: bot, provider,
    persona, im, config, chat, kb, memory, data, file, plugin, mcp, skill. The
    config scope also grants bot and provider access.
servers:
  - url: http://localhost:6185
    description: Local AstrBot server
security:
  - ApiKeyAuth: []
...
paths:
...
  /api/v1/files:
    post:
      tags: [Files]
      summary: Upload a file
      operationId: uploadFile
      x-astrbot-scope: file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/FileUploadRequest'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
...
  /api/v1/plugins:
    get:
      tags: [Plugins]
      summary: List installed plugins
      operationId: listPlugins
      x-astrbot-scope: plugin
      parameters:
        - name: include_reserved
          in: query
          schema:
            type: boolean
            default: true
        - name: enabled
          in: query
          schema:
            type: boolean
      responses:
        '200':
          $ref: '#/components/responses/Ok'
  /api/v1/plugins/{plugin_id}:
    get:
      tags: [Plugins]
      summary: Get plugin details
      operationId: getPlugin
      x-astrbot-scope: plugin
      parameters:
        - $ref: '#/components/parameters/PluginId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                delete_config:
                  type: boolean
                delete_data:
                  type: boolean
              additionalProperties: true
      responses:
        '200':
          $ref: '#/components/responses/Ok'
    delete:
      tags: [Plugins]
      summary: Uninstall a plugin
      operationId: uninstallPlugin
      x-astrbot-scope: plugin
      parameters:
        - $ref: '#/components/parameters/PluginId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                delete_config:
                  type: boolean
                delete_data:
                  type: boolean
              additionalProperties: true
      responses:
        '200':
          $ref: '#/components/responses/Ok'
  /api/v1/plugins/{plugin_id}/config:
    get:
      tags: [Plugins]
      summary: Get plugin configuration
      operationId: getPluginConfig
      x-astrbot-scope: plugin
      parameters:
        - $ref: '#/components/parameters/PluginId'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
    put:
      tags: [Plugins]
      summary: Save plugin configuration
      operationId: updatePluginConfig
      x-astrbot-scope: plugin
      parameters:
        - $ref: '#/components/parameters/PluginId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DynamicConfig'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
  /api/v1/plugins/{plugin_id}/config/schema:
    get:
      tags: [Plugins]
      summary: Get plugin configuration schema
      operationId: getPluginConfigSchema
      x-astrbot-scope: plugin
      parameters:
        - $ref: '#/components/parameters/PluginId'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
  /api/v1/plugins/{plugin_id}/config-files/{config_key}:
    get:
      tags: [Plugins]
      summary: List uploaded files for a plugin file configuration item
      operationId: listPluginConfigFiles
      x-astrbot-scope: plugin
      parameters:
        - $ref: '#/components/parameters/PluginId'
        - name: config_key
          in: path
          required: true
          description: URL-encoded configuration key path.
          schema:
            type: string
      responses:
        '200':
          $ref: '#/components/responses/Ok'
    post:
      tags: [Plugins]
      summary: Upload files for a plugin file configuration item
      operationId: uploadPluginConfigFiles
      x-astrbot-scope: plugin
      parameters:
        - $ref: '#/components/parameters/PluginId'
        - name: config_key
          in: path
          required: true
          description: URL-encoded configuration key path.
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          $ref: '#/components/responses/Ok'
  /api/v1/plugins/{plugin_id}/config-files:
    delete:
      tags: [Plugins]
      summary: Delete a plugin file configuration upload
      operationId: deletePluginConfigFile
      x-astrbot-scope: plugin
      parameters:
        - $ref: '#/components/parameters/PluginId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PluginConfigFileDeleteRequest'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
  /api/v1/plugins/{plugin_id}/readme:
    get:
      tags: [Plugins]
      summary: Get plugin README content
      operationId: getPluginReadme
      x-astrbot-scope: plugin
      parameters:
        - $ref: '#/components/parameters/PluginId'
      responses:
        '200':
          $ref: '#/components/responses/Text'
  /api/v1/plugins/{plugin_id}/changelog:
    get:
      tags: [Plugins]
      summary: Get plugin CHANGELOG content
      operationId: getPluginChangelog
      x-astrbot-scope: plugin
      parameters:
        - $ref: '#/components/parameters/PluginId'
      responses:
        '200':
          $ref: '#/components/responses/Text'
  /api/v1/plugins/{plugin_id}/reload:
    post:
      tags: [Plugins]
      summary: Reload a plugin
      operationId: reloadPlugin
      x-astrbot-scope: plugin
      parameters:
        - $ref: '#/components/parameters/PluginId'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
  /api/v1/plugins/{plugin_id}/source:
    post:
      tags: [Plugins]
      summary: Bind an installed plugin to a marketplace source
      operationId: bindPluginSource
      x-astrbot-scope: plugin
      parameters:
        - $ref: '#/components/parameters/PluginId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PluginSourceBindRequest'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
  /api/v1/plugins/{plugin_id}/enabled:
    patch:
      tags: [Plugins]
      summary: Enable or disable a plugin
      operationId: setPluginEnabled
      x-astrbot-scope: plugin
      parameters:
        - $ref: '#/components/parameters/PluginId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnabledPatch'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
  /api/v1/plugins/{plugin_id}/update:
    post:
      tags: [Plugins]
      summary: Update a plugin
      operationId: updatePlugin
      x-astrbot-scope: plugin
      parameters:
        - $ref: '#/components/parameters/PluginId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PluginUpdateRequest'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
  /api/v1/plugins/update:
    post:
      tags: [Plugins]
      summary: Update multiple plugins
      operationId: updatePlugins
      x-astrbot-scope: plugin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PluginBatchUpdateRequest'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
...
  /api/v1/plugins/failed/{plugin_id}:
    delete:
      tags: [Plugins]
      summary: Uninstall a failed plugin
      operationId: uninstallFailedPlugin
      x-astrbot-scope: plugin
      parameters:
        - $ref: '#/components/parameters/PluginId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                delete_config:
                  type: boolean
                delete_data:
                  type: boolean
              additionalProperties: true
      responses:
        '200':
          $ref: '#/components/responses/Ok'
  /api/v1/plugins/failed/{plugin_id}/reload:
    post:
      tags: [Plugins]
      summary: Reload a failed plugin
      operationId: reloadFailedPlugin
      x-astrbot-scope: plugin
      parameters:
        - $ref: '#/components/parameters/PluginId'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
  /api/v1/plugins/install/github:
    post:
      tags: [Plugins]
      summary: Install a plugin from GitHub
      operationId: installPluginFromGithub
      x-astrbot-scope: plugin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PluginGithubInstallRequest'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
  /api/v1/plugins/install/url:
    post:
      tags: [Plugins]
      summary: Install a plugin from a downloadable ZIP URL
      operationId: installPluginFromUrl
      x-astrbot-scope: plugin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PluginUrlInstallRequest'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
  /api/v1/plugins/install/upload:
    post:
      tags: [Plugins]
      summary: Install a plugin from an uploaded ZIP file
      operationId: installPluginFromUpload
      x-astrbot-scope: plugin
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PluginUploadInstallRequest'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
  /api/v1/plugins/market:
    get:
      tags: [Plugins]
      summary: List plugin marketplace entries
      operationId: listPluginMarket
      x-astrbot-scope: plugin
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: category
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [recommended, downloads, updated, name]
        - name: keyword
          in: query
          schema:
            type: string
        - name: force_refresh
          in: query
          schema:
            type: boolean
        - name: custom_registry
          in: query
          schema:
            type: string
      responses:
        '200':
          $ref: '#/components/responses/Ok'
...
  /api/v1/plugin-sources:
    get:
      tags: [Plugin Sources]
      summary: List plugin sources
      operationId: listPluginSources
      x-astrbot-scope: plugin
      responses:
        '200':
          $ref: '#/components/responses/Ok'
    post:
      tags: [Plugin Sources]
      summary: Add a plugin source
      operationId: createPluginSource
      x-astrbot-scope: plugin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PluginSourceRequest'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
    put:
      tags: [Plugin Sources]
      summary: Replace plugin sources
      operationId: replacePluginSources
      x-astrbot-scope: plugin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sources]
              properties:
                sources:
                  type: array
                  items:
                    $ref: '#/components/schemas/PluginSourceRequest'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
  /api/v1/plugin-sources/{source_id}:
    delete:
      tags: [Plugin Sources]
      summary: Delete a plugin source
      operationId: deletePluginSource
      x-astrbot-scope: plugin
      parameters:
        - $ref: '#/components/parameters/SourceId'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    DashboardBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Dashboard session JWT
    DashboardCookieAuth:
      type: apiKey
      in: cookie
      name: astrbot_dashboard_jwt
  responses:
    Ok:
      description: Standard AstrBot success response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SuccessEnvelope'
...
    Text:
      description: Text response
      content:
        text/plain:
          schema:
            type: string
        application/json:
          schema:
            $ref: '#/components/schemas/SuccessEnvelope'
...
  parameters:
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 20
...
    PluginId:
      name: plugin_id
      in: path
      required: true
      schema:
        type: string
...
    SourceId:
      name: source_id
      in: path
      required: true
      schema:
        type: string
...
  schemas:
    SuccessEnvelope:
      type: object
      required: [status, data]
      properties:
        status:
          type: string
          const: ok
        message:
          type: string
        data: {}
      additionalProperties: false
...
    DynamicConfig:
      type: object
      additionalProperties: true
...
    EnabledPatch:
      type: object
      required: [enabled]
      properties:
        enabled:
          type: boolean
      additionalProperties: false
...
    FileUploadRequest:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
    PluginUpdateRequest:
      type: object
      additionalProperties: false
      properties:
        proxy:
          type: string
          nullable: true
    PluginSourceBindRequest:
      type: object
      properties:
        install_method:
          type: string
        registry_url:
          type: string
          nullable: true
        market_plugin_id:
          type: string
      additionalProperties: false
    PluginBatchUpdateRequest:
      type: object
      required: [names]
      additionalProperties: false
      properties:
        names:
          type: array
          items:
            type: string
        proxy:
          type: string
          nullable: true
    PluginGithubInstallRequest:
      type: object
      required: [repository]
      properties:
        repository:
          type: string
          description: GitHub URL or owner/repository slug.
        download_url:
          type: string
          description: Optional downloadable ZIP URL to use instead of GitHub archive resolution.
        proxy:
          type: string
        ignore_version_check:
          type: boolean
        install_method:
          type: string
        registry_url:
          type: string
          nullable: true
        market_plugin_id:
          type: string
      additionalProperties: false
    PluginUrlInstallRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
        download_url:
          type: string
          format: uri
          description: Optional downloadable ZIP URL when url is the plugin source page or repository.
        proxy:
          type: string
        ignore_version_check:
          type: boolean
        install_method:
          type: string
        registry_url:
          type: string
          nullable: true
        market_plugin_id:
          type: string
      additionalProperties: false
    PluginUploadInstallRequest:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
    PluginConfigFileDeleteRequest:
      type: object
      required: [path]
      properties:
        path:
          type: string
      additionalProperties: false
    PluginSourceRequest:
      type: object
      required: [url]
      properties:
        id:
          type: string
        name:
          type: string
        url:
          type: string
          format: uri
      additionalProperties: false
...